+++ ui/src/kapacitor/actions/view/index.js
@@ -1,7 +1,12 @@
 import uuid from 'node-uuid';
-import {getRules, getRule, deleteRule as deleteRuleAPI} from 'src/kapacitor/apis';
 import {getKapacitor} from 'src/shared/apis';
 import {publishNotification} from 'src/shared/actions/notifications';
+import {
+  getRules,
+  getRule,
+  deleteRule as deleteRuleAPI,
+  updateRuleStatus as updateRuleStatusAPI,
+} from 'src/kapacitor/apis';
 
 export function fetchRule(source, ruleID) {
   return (dispatch) => {
@@ -37,7 +42,7 @@ export function loadDefaultRule() {
     dispatch({
       type: 'ADD_KAPACITOR_QUERY',
       payload: {
-        queryId: queryID,
+        queryID,
       },
     });
   };
@@ -87,12 +92,33 @@ export function updateMessage(ruleID, message) {
   };
 }
 
+export function updateDetails(ruleID, details) {
+  return {
+    type: 'UPDATE_RULE_DETAILS',
+    payload: {
+      ruleID,
+      details,
+    },
+  };
+}
+
 export function updateAlerts(ruleID, alerts) {
   return {
     type: 'UPDATE_RULE_ALERTS',
     payload: {
       ruleID,
       alerts,
+    },
+  };
+}
+
+export function updateAlertNodes(ruleID, alertType, alertNodesText) {
+  return {
+    type: 'UPDATE_RULE_ALERT_NODES',
+    payload: {
+      ruleID,
+      alertType,
+      alertNodesText,
     },
   };
 }
@@ -116,6 +142,16 @@ export function deleteRuleSuccess(ruleID) {
   };
 }
 
+export function updateRuleStatusSuccess(ruleID, status) {
+  return {
+    type: 'UPDATE_RULE_STATUS_SUCCESS',
+    payload: {
+      ruleID,
+      status,
+    },
+  };
+}
+
 export function deleteRule(rule) {
   return (dispatch) => {
     deleteRuleAPI(rule).then(() => {
@@ -123,6 +159,17 @@ export function deleteRule(rule) {
       dispatch(publishNotification('success', `${rule.name} deleted successfully`));
     }).catch(() => {
       dispatch(publishNotification('error', `${rule.name} could not be deleted`));
+    });
+  };
+}
+
+export function updateRuleStatus(rule, {status}) {
+  return (dispatch) => {
+    updateRuleStatusAPI(rule, status).then(() => {
+      dispatch(publishNotification('success', `${rule.name} ${status} successfully`));
+    }).catch(() => {
+      dispatch(updateRuleStatusSuccess(rule.id, status));
+      dispatch(publishNotification('error', `${rule.name} could not be ${status}`));
     });
   };
 }